Makefile now calls fontforge-embedded python
When homebrew installs fontforge, it's supposed to both:
- embed python into fontforge and 
- install fontforge as a python extension
For my installation, the latter step failed. Looking at the issues on the fontforge project, I found several people with the same problem on Mac, but no simple fix. The workaround is to instead call the version of python embedded into fontforge, which can be done using `fontforge -script foo.py {arguments}`